Update barClass to return a boolean
You can only use one ng-class directive. When there are multiple, only the first is used. Instead
you have to list you classes together like this `ng-class="{'active': uploading,
'progress-bar-success': barClass()}"`.

This commit converts the `barClass` fn to return boolean instead of an object so the value can be
used in both the bootstrap2 template and the bootstrap3 template - where it is combined with the
`active` class.